chore: Refactor names of buttons and add brief informations
