+++ src/buttons/index.ts
@@ -1,2 +1,2 @@
-export * as addChannel from './addChannel';
-export * as openCreateModal from './openCreateModal';
+export * as addChannel from './addChannelButton';
+export * as openCreateModal from './openCreateModalButton';
