+++ src/commands/index.ts
@@ -7,4 +7,4 @@ export * as announce from './announce';
 export * as restart from './restart';
 export * as verify from './verify';
 export * as config from './config';
-export * as channelmanagement from './channelAssign';
+export * as channelmanagement from './channelmanagement';
